Add files via upload
This is a Hostel Management System. There is Head of hostel or warden who can add or reduce people in the hostel. There is also a form for student. And they can select their rooms . A complete project which can help you to have an understanding on working with java in an easy way. It contains the use of NetBeans as IDE, Html for look and feel of project, CSS for presentation, SQL as Querry language.